@@ -4,7 +4,7 @@ version: 0.0.7
 repository: https://github.com/joranmulderij/flutter_guitar_tabs
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
